web/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTX==1.3.2
requests==2.31.0
PyJWT==2.8.0
redis==4.5.4